refactor: use NodejsFunction instead of lambda.Function
