--- server/handlers/index.js
@@ -1,9 +1,0 @@
-const { getProductsList } = require("./products/getProductsList");
-const { getProductById } = require("./products/getProductById");
-const { createProduct } = require("./products/createProduct");
-
-module.exports = {
-  getProductsList,
-  getProductById,
-  createProduct,
-};
