+++ server/handlers/products/index.js
@@ -1,0 +1,9 @@
+const { createProduct } = require("./createProduct");
+const { getProductById } = require("./getProductById");
+const { getProductsList } = require("./getProductsList");
+
+module.exports = {
+  createProduct,
+  getProductById,
+  getProductsList,
+};
